Move summariseMatching to tests/helpers/comparison folder
