+++ packages/core/phase1/phase1.comparison.test.ts
@@ -10,7 +10,6 @@ import type { OldPhase1Comparison } from "../tests/types/ComparisonFile"
 import type Exception from "../types/Exception"
 import type Phase1Result from "./types/Phase1Result"
 
-import { matchingExceptions } from "../comparison/lib/summariseMatching"
 import CoreAuditLogger from "../lib/auditLog/CoreAuditLogger"
 import saveErrorListRecord from "../lib/database/saveErrorListRecord"
 import { extractExceptionsFromXml } from "../lib/parse/parseAhoXml"
@@ -21,6 +20,7 @@ import generateMockPncQueryResultFromAho from "../tests/helpers/comparison/gener
 import getComparisonTests from "../tests/helpers/comparison/getComparisonTests"
 import getPncQueryTimeFromAho from "../tests/helpers/comparison/getPncQueryTimeFromAho"
 import parseIncomingMessage from "../tests/helpers/comparison/parseIncomingMessage"
+import { matchingExceptions } from "../tests/helpers/comparison/summariseMatching"
 import MockPncGateway from "../tests/helpers/MockPncGateway"
 import sortExceptions from "../tests/helpers/sortExceptions"
 import phase1 from "./phase1"
